% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddy-package.R
\docType{package}
\name{buddy-package}
\alias{buddy}
\alias{buddy-package}
\title{buddy: Package for Buddy's names}
\description{
Package to test Buddy's nicknames
}
\author{
\strong{Maintainer}: Asmae Toumi \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
