% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_buddy.R
\name{is_buddy}
\alias{is_buddy}
\title{Check for Buddy's nicknames}
\usage{
is_buddy(name)
}
\arguments{
\item{name}{A nickname of Buddy;s.}
}
\description{
Check whether a name is indeed Buddy's nickname
}
\examples{
is_buddy("bud")
}
